Refactored add waveform logic in session.
